extract getHoursOfOperationRange into client/utils
